// src/app/perfil/perfil.enum.ts
export enum Perfil {
    Professor = 1,
    Administrador = 2,
    Curador = 4,
    Produtora = 5,
    GestorConteudoIes = 7,
    GestorComercialIes = 8,
    CoordenadorAcademico = 9,
    FinanceiroIes = 10,
    MarketingIes = 11,
    AtendimentoIes = 12,
    SecretariaIes = 13,
    Aluno = 14,
    Projeto = 15,
    TesteValidacao = 16,
    Internauta = 17,
    OperadorIes = 18,
    Parceiro = 19,
    Desenvolvedor = 20,
    SecretariaEducaz = 21,
    GestorIes = 22,
}